old-tests/submission/Infraware/Forms/contents/Forms/keygen_challenge.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>keygen_challenge</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <p><keygen name="key" id='keygen_id'></keygen></p>
	</form>

	<script>
		
		var keygen = document.getElementById("keygen_id");
		keygen.challenge = 'challenge';

		if (typeof(keygen.challenge) == "string") {
			test(function() {
				assert_equals(keygen.challenge, "challenge", "challenge attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("challenge attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
